added journal rcView stuff

+++ app/src/main/java/com/example/luigi/journally/LocationModel.java
@@ -2,13 +2,13 @@ package com.example.luigi.journally;
 
 public class LocationModel {
 
-    private long id;
+    private int id;
     private String name;
     private double lat;
     private double longt;
     private String timestamp;
 
-    public LocationModel(long id, String name, double lat, double longt, String timestamp)
+    public LocationModel(int id, String name, double lat, double longt, String timestamp)
     {
         this.setTimestamp(timestamp);
         this.setId(id);
@@ -41,11 +41,11 @@ public class LocationModel {
         this.longt = longt;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
